requirements.txt
openai
python-dotenv
colorama
time
virtualenv
openpyxl
pandas
csv